fix release + tests
need to separate test runner from bin to be releasable
