+++ .github/workflows/ci.yml
@@ -24,11 +24,10 @@ jobs:
           github-token: ${{ secrets.GITHUB_TOKEN }}
       - run: kubectl apply --force-conflicts --server-side -f https://raw.githubusercontent.com/prometheus-operator/prometheus-operator/v0.52.0/example/prometheus-operator-crd/monitoring.coreos.com_prometheusrules.yaml
       - run: cargo build --bin kopium
-      - run: cargo run --bin kopium -- prometheusrules.monitoring.coreos.com > gen.rs
-      - run: echo "pub type CR = PrometheusRule;" >> gen.rs
-      - run: kubectl apply -f test.yaml
-      - run: cargo run --bin tester -- PrometheusRule
-
+      - run: cargo run --bin kopium -- prometheusrules.monitoring.coreos.com > tests/gen.rs
+      - run: echo "pub type CR = PrometheusRule;" >> tests/gen.rs
+      - run: kubectl apply -f tests/pr.yaml
+      - run: cargo test --test runner -- --nocapture
 
   rustfmt:
     name: Run rustfmt
